1816 collections wrong counts when creating collection (#1820)
* Improve filter summary formatting and operator labels

Updated filter summary to display 'AND' between filters and use small-caps styling for operators. Changed operator translations to uppercase ('AND', 'OR') for consistency. Adjusted summary text handling and improved input description formatting in the search page. Fix #1816

* Refactor AbortController usage in DataRundown.vue

Moves the AbortController to a ref at the component level to ensure it is accessible in both onMounted and onBeforeUnmount hooks. This change improves request cancellation handling when the component is unmounted.

---------

Co-authored-by: Daniele Guido <1181642+danieleguido@users.noreply.github.com>